Configure global logging substitution via `config!`
- Added `substitute` keyword argument to `config!`
- Added appropriate notice message (and warning when substitution fails on 0.6)
- Added some test cases.
